Move import ...traceability.model.DockerTraceabilityReport to ..traceability.api.DockerTraceabilityReport
