+++ docker-traceability-plugin/src/main/java/org/jenkinsci/plugins/docker/traceability/core/DockerTraceabilityRootAction.java
@@ -58,6 +58,7 @@ import jenkins.model.Jenkins;
 import static jenkins.model.Jenkins.XSTREAM;
 import org.acegisecurity.AccessDeniedException;
 import org.apache.commons.lang.StringUtils;
+import org.jenkinsci.plugins.docker.traceability.api.DockerTraceabilityReport;
 import org.jenkinsci.plugins.docker.traceability.model.DockerTraceabilityReportListener;
 import org.jenkinsci.plugins.docker.traceability.DockerTraceabilityPlugin;
 import org.jenkinsci.plugins.docker.traceability.dockerjava.api.command.InspectContainerResponse;
@@ -70,7 +71,6 @@ import org.jenkinsci.plugins.docker.traceability.model.DockerAPIReport;
 import org.jenkinsci.plugins.docker.traceability.model.DockerEvent;
 import org.jenkinsci.plugins.docker.traceability.model.DockerEventType;
 import org.jenkinsci.plugins.docker.traceability.model.DockerInfo;
-import org.jenkinsci.plugins.docker.traceability.model.DockerTraceabilityReport;
 import org.jenkinsci.plugins.docker.traceability.util.FingerprintsHelper;
 import org.kohsuke.accmod.Restricted;
 import org.kohsuke.accmod.restrictions.NoExternalUse;
